Refactor imports and reorganize tool factory structure
