+++ dev_kit_mcp_server/fastmcp_server.py
@@ -9,9 +9,9 @@ from pathlib import Path
 # from mcp.server.fastmcp import FastMCP  # type: ignore
 from fastmcp import FastMCP
 
+from .tool_factory import ToolFactory
 from .tools import CreateDirOperation, MoveDirOperation, RemoveFileOperation, RenameOperation
 from .tools.commands_tool import ExecMakeTarget
-from .tools.tool_factory import ToolFactory
 
 
 def start_server(root_dir: str = None) -> FastMCP:
