+++ dev_kit_mcp_server/tools/__init__.py
@@ -6,16 +6,7 @@ from .move import MoveDirOperation
 from .remove import RemoveFileOperation
 from .rename import RenameOperation
 
-# Import from code_editing
-# Import from tool factory
-from .tool_factory import ToolFactory
-
-# Import utilities
-
 __all__ = [
-    # Code editing tools
-    # Tool factory
-    "ToolFactory",
     "CreateDirOperation",
     "RemoveFileOperation",
     "MoveDirOperation",
